[HOTSPOT-188] fix: refactoring 오류 수정
